Fix test_consumer test failures
