+++ tests/integration/wsapi/test_consumer.py
@@ -13,6 +13,10 @@ from aap_eda.core import models
 from aap_eda.core.enums import CredentialType
 from aap_eda.wsapi.consumers import AnsibleRulebookConsumer
 
+# TODO(doston): this test module needs a whole refactor to use already
+# existing fixtures over from API conftest.py instead of creating new objects
+# keeping it to minimum for now to pass all failing tests
+
 TIMEOUT = 5
 
 DATETIME_FORMAT = "%Y-%m-%dT%H:%M:%S.%fZ"
@@ -41,6 +45,7 @@ DUMMY_UUID2 = "8472ff2c-6045-4418-8d4e-46f6cfffffff"
 @pytest.mark.django_db(transaction=True)
 async def test_handle_workers_without_credentials(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_with_extra_var = await _prepare_db_data()
     rulebook_process_without_extra_var = (
@@ -98,6 +103,7 @@ async def test_handle_workers_without_credentials(
 @pytest.mark.django_db(transaction=True)
 async def test_handle_workers_with_system_vault_credential(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = (
         await _prepare_activation_instance_with_system_vault_credential()
@@ -127,6 +133,7 @@ async def test_handle_workers_with_system_vault_credential(
 @pytest.mark.django_db(transaction=True)
 async def test_handle_workers_with_vault_credentials(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = (
         await _prepare_activation_instance_with_vault_credentials()
@@ -156,6 +163,7 @@ async def test_handle_workers_with_vault_credentials(
 @pytest.mark.django_db(transaction=True)
 async def test_handle_workers_with_all_credentials(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = (
         await _prepare_activation_instance_with_all_credentials()
@@ -188,7 +196,9 @@ async def test_handle_workers_with_all_credentials(
 
 
 @pytest.mark.django_db(transaction=True)
-async def test_handle_workers_with_validation_errors():
+async def test_handle_workers_with_validation_errors(
+    default_organization: models.Organization,
+):
     communicator = WebsocketCommunicator(
         AnsibleRulebookConsumer.as_asgi(), "ws/"
     )
@@ -208,7 +218,10 @@ async def test_handle_workers_with_validation_errors():
 
 
 @pytest.mark.django_db(transaction=True)
-async def test_handle_jobs(ws_communicator: WebsocketCommunicator):
+async def test_handle_jobs(
+    ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
+):
     rulebook_process_id = await _prepare_db_data()
 
     assert (await get_job_instance_count()) == 0
@@ -255,6 +268,7 @@ async def test_handle_events(ws_communicator: WebsocketCommunicator):
 @pytest.mark.django_db(transaction=True)
 async def test_handle_actions_multiple_firing(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = await _prepare_db_data()
     job_instance = await _prepare_job_instance()
@@ -288,6 +302,7 @@ async def test_handle_actions_multiple_firing(
 @pytest.mark.django_db(transaction=True)
 async def test_handle_actions_with_empty_job_uuid(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = await _prepare_db_data()
     assert (await get_audit_rule_count()) == 0
@@ -310,7 +325,10 @@ async def test_handle_actions_with_empty_job_uuid(
 
 
 @pytest.mark.django_db(transaction=True)
-async def test_handle_actions(ws_communicator: WebsocketCommunicator):
+async def test_handle_actions(
+    ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
+):
     rulebook_process_id = await _prepare_db_data()
     job_instance = await _prepare_job_instance()
 
@@ -352,6 +370,7 @@ async def test_handle_actions(ws_communicator: WebsocketCommunicator):
 @pytest.mark.django_db(transaction=True)
 async def test_rule_status_with_multiple_failed_actions(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = await _prepare_db_data()
     job_instance = await _prepare_job_instance()
@@ -385,7 +404,10 @@ async def test_rule_status_with_multiple_failed_actions(
 
 
 @pytest.mark.django_db(transaction=True)
-async def test_handle_heartbeat(ws_communicator: WebsocketCommunicator):
+async def test_handle_heartbeat(
+    ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
+):
     rulebook_process_id = await _prepare_db_data()
     rulebook_process = await get_rulebook_process(rulebook_process_id)
 
@@ -422,6 +444,7 @@ async def test_handle_heartbeat(ws_communicator: WebsocketCommunicator):
 @pytest.mark.django_db(transaction=True)
 async def test_multiple_rules_for_one_event(
     ws_communicator: WebsocketCommunicator,
+    default_organization: models.Organization,
 ):
     rulebook_process_id = await _prepare_db_data()
     job_instance = await _prepare_job_instance()
